demo: Moved svelte components to common location for demo and test-viewer
