+++ src/components/Editor.svelte
