+++ src/components/Graph.svelte
@@ -1,20 +1,20 @@
 <script lang="ts">
   import Parser from "web-tree-sitter";
-  import { newCFGBuilder, type Language } from "../../../control-flow/cfg";
+  import { newCFGBuilder, type Language } from "../control-flow/cfg";
   import {
     mergeNodeAttrs,
     remapNodeTargets,
     type CFG,
-  } from "../../../control-flow/cfg-defs";
-  import { graphToDot } from "../../../control-flow/render";
-  import { simplifyCFG, trimFor } from "../../../control-flow/graph-ops";
+  } from "../control-flow/cfg-defs";
+  import { graphToDot } from "../control-flow/render";
+  import { simplifyCFG, trimFor } from "../control-flow/graph-ops";
   import { Graphviz } from "@hpcc-js/wasm-graphviz";
   import {
     getFirstFunction,
     initialize as initializeUtils,
     type Parsers,
   } from "./utils";
-  import { getValue } from "../../../control-flow/ranges";
+  import { getValue } from "../control-flow/ranges";
   import { createEventDispatcher } from "svelte";
 
   let parsers: Parsers;
