+++ src/frontend/src/App.svelte
@@ -1,5 +1,5 @@
 <script>
-  import TestViewer from "./lib/TestViewer.svelte";
+  import TestViewer from "../../components/TestViewer.svelte";
 </script>
 
 <main>
