+++ src/vscode/extension.ts
@@ -288,6 +288,6 @@ export async function activate(context: vscode.ExtensionContext) {
 }
 
 // This method is called when your extension is deactivated
-export function deactivate() { }
+export function deactivate() {}
 
 //------------------------------------------------
